Refactoring funzione che permette di ottenere i suoi elementi adiacenti data una coordinata.
La Funzione che permette di ottenere gli elementi data una coordinata è presumibilmente una responsabilità della board pertanto è stato spostata nella classe Board.java insieme al relativo test ed è stato effettuato un refactoring sulla stessa.
